moved test artifacts to other folder
