+++ tests/BoltOn.Tests/BoltOn.Tests.csproj
@@ -19,6 +19,7 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="2.2.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -30,6 +31,7 @@
     <Folder Include="Data\EF\" />
     <Folder Include="Mediator\Data\" />
     <Folder Include="Mediator\Data\EF\" />
+    <Folder Include="Other\" />
   </ItemGroup>
 
   <ItemGroup>
