+++ tests/BoltOn.Tests/Other/TestDataRegistrationTask.cs
@@ -4,18 +4,24 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace BoltOn.Tests.Data.EF
+namespace BoltOn.Tests.Other
 {
 	public class TestDataRegistrationTask : IBootstrapperRegistrationTask
 	{
 		public void Run(RegistrationTaskContext context)
 		{
-			if (MediatorTestHelper.IsSqlite)
+			if (MediatorTestHelper.IsSqlServer)
 			{
+				//context.Container.AddDbContext<SchoolDbContext>(options =>
+				//{
+				//	options.UseSqlite("Data Source=TestDatabase.db");
+				//	options.ConfigureWarnings(x => x.Ignore(RelationalEventId.AmbientTransactionWarning));
+				//});
+
+
 				context.Container.AddDbContext<SchoolDbContext>(options =>
 				{
-					options.UseSqlite("Data Source=TestDatabase.db");
-					options.ConfigureWarnings(x => x.Ignore(RelationalEventId.AmbientTransactionWarning));
+					options.UseSqlServer("Data Source=127.0.0.1;initial catalog=Testing;persist security info=True;User ID=sa;Password=$Password1;");
 				});
 			}
 			else
