+++ tests/BoltOn.Tests/Other/TestInterceptor.cs
@@ -5,7 +5,7 @@ using BoltOn.Logging;
 using BoltOn.Mediator.Interceptors;
 using BoltOn.Mediator.Pipeline;
 
-namespace BoltOn.Tests.Mediator
+namespace BoltOn.Tests.Other
 {
 	public class TestInterceptor : IInterceptor
 	{
