+++ tests/BoltOn.Tests/Other/TestMediatorRegistrationTask.cs
@@ -3,12 +3,13 @@ using BoltOn.Bootstrapping;
 using BoltOn.Logging;
 using BoltOn.Mediator;
 using BoltOn.Mediator.Interceptors;
+using BoltOn.Tests.Mediator;
 using BoltOn.UoW;
 using BoltOn.Utilities;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
 
-namespace BoltOn.Tests.Mediator
+namespace BoltOn.Tests.Other
 {
 	public class TestMediatorRegistrationTask : IBootstrapperRegistrationTask
 	{
